# pkg/api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Daytona Server API",
        "title": "Daytona Server API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:3986",
    "basePath": "/",
    "paths": {
...
        "/project-config": {
            "get": {
                "description": "List project configs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-config"
                ],
                "summary": "List project configs",
                "operationId": "ListProjectConfigs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ProjectConfig"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Set project config data",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "project-config"
                ],
                "summary": "Set project config data",
                "operationId": "SetProjectConfig",
                "parameters": [
                    {
                        "description": "Project config",
                        "name": "projectConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateProjectConfigDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/project-config/{configName}": {
            "get": {
                "description": "Get project config data",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "project-config"
                ],
                "summary": "Get project config data",
                "operationId": "GetProjectConfig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Config name",
                        "name": "configName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ProjectConfig"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete project config data",
                "tags": [
                    "project-config"
                ],
                "summary": "Delete project config data",
                "operationId": "DeleteProjectConfig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Config name",
                        "name": "configName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/project-config/{gitUrl}": {
            "get": {
                "description": "Get project configs by git url",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-config"
                ],
                "summary": "Get project configs by git url",
                "operationId": "GetDefaultProjectConfig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Git URL",
                        "name": "gitUrl",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ProjectConfig"
                        }
                    }
                }
            }
        },
...
        "/workspace": {
            "get": {
                "description": "List workspaces",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "List workspaces",
                "operationId": "ListWorkspaces",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Verbose",
                        "name": "verbose",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/WorkspaceDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a workspace",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Create a workspace",
                "operationId": "CreateWorkspace",
                "parameters": [
                    {
                        "description": "Create workspace",
                        "name": "workspace",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateWorkspaceDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Workspace"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateProjectConfigDTO": {
            "type": "object",
            "properties": {
                "build": {
                    "$ref": "#/definitions/ProjectBuild"
                },
                "envVars": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "source": {
                    "$ref": "#/definitions/CreateProjectConfigSourceDTO"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "CreateProjectConfigSourceDTO": {
            "type": "object",
            "properties": {
                "repository": {
                    "$ref": "#/definitions/GitRepository"
                }
            }
        },
        "CreateProjectDTO": {
            "type": "object",
            "properties": {
                "existingProjectConfig": {
                    "$ref": "#/definitions/ExistingProjectConfigDTO"
                },
                "newProjectConfig": {
                    "$ref": "#/definitions/CreateProjectConfigDTO"
                }
            }
        },
        "CreateWorkspaceDTO": {
            "type": "object",
            "required": [
                "projects"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CreateProjectDTO"
                    }
                },
                "target": {
                    "type": "string"
                }
            }
        },
        "ExistingProjectConfigDTO": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "configName": {
                    "type": "string"
                },
                "projectName": {
                    "type": "string"
                }
            }
        },
...
        "FileStatus": {
            "type": "object",
            "properties": {
                "extra": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "staging": {
                    "$ref": "#/definitions/Status"
                },
                "worktree": {
                    "$ref": "#/definitions/Status"
                }
            }
        },
...
        "GitRepository": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "prNumber": {
                    "type": "integer"
                },
                "sha": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "GitStatus": {
            "type": "object",
            "properties": {
                "currentBranch": {
                    "type": "string"
                },
                "fileStatus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/FileStatus"
                    }
                }
            }
        },
...
        "Project": {
            "type": "object",
            "properties": {
                "build": {
                    "$ref": "#/definitions/ProjectBuild"
                },
                "default": {
                    "type": "boolean"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "repository": {
                    "$ref": "#/definitions/GitRepository"
                },
                "state": {
                    "$ref": "#/definitions/ProjectState"
                },
                "target": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                },
                "workspaceId": {
                    "type": "string"
                }
            }
        },
        "ProjectBuild": {
            "type": "object",
            "properties": {
                "devcontainer": {
                    "$ref": "#/definitions/ProjectBuildDevcontainer"
                }
            }
        },
        "ProjectBuildDevcontainer": {
            "type": "object",
            "properties": {
                "devContainerFilePath": {
                    "type": "string"
                }
            }
        },
        "ProjectConfig": {
            "type": "object",
            "properties": {
                "build": {
                    "$ref": "#/definitions/ProjectBuild"
                },
                "default": {
                    "type": "boolean"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "repository": {
                    "$ref": "#/definitions/GitRepository"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "ProjectInfo": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "isRunning": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "providerMetadata": {
                    "type": "string"
                },
                "workspaceId": {
                    "type": "string"
                }
            }
        },
        "ProjectState": {
            "type": "object",
            "properties": {
                "gitStatus": {
                    "$ref": "#/definitions/GitStatus"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uptime": {
                    "type": "integer"
                }
            }
        },
...
        "Status": {
            "type": "string",
            "enum": [
                "Unmodified",
                "Untracked",
                "Modified",
                "Added",
                "Deleted",
                "Renamed",
                "Copied",
                "Updated but unmerged"
            ],
            "x-enum-varnames": [
                "Unmodified",
                "Untracked",
                "Modified",
                "Added",
                "Deleted",
                "Renamed",
                "Copied",
                "UpdatedButUnmerged"
            ]
        },
        "Workspace": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Project"
                    }
                },
                "target": {
                    "type": "string"
                }
            }
        },
        "WorkspaceDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "info": {
                    "$ref": "#/definitions/WorkspaceInfo"
                },
                "name": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Project"
                    }
                },
                "target": {
                    "type": "string"
                }
            }
        },
        "WorkspaceInfo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ProjectInfo"
                    }
                },
                "providerMetadata": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}